+++ docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v.0.1.1
+
+### Hinzugefügt
+
+- Installationsskript (`install.sh`) zum automatischen Erstellen einer Python-Virtual-Environment und Kompilieren des Projekts.
+- `requirements.txt` zur Verwaltung der Python-Abhängigkeiten.
+- Installationsanleitung in der `README.md` ergänzt.
+- Kleiner Fix: alias "initialize_from_address" in module.cpp ergänzt.
+
 ## v0.1.0
 
 ### Hinzugefügt